+++ js-miniapp-bridge/src/types/universal-bridge.ts
@@ -1,0 +1,8 @@
+/**
+ * Model that will be sent to Host app through Universal bridge
+ */
+export interface UniversalBridgeInfo {
+  key?: string;
+  value?: string;
+  description?: string;
+}